Add files via upload
Updated with the start of queries (ones that are not completed are marked "NOT COMPLETED"). I can't test to see if they work because my sqldeveloper freaks out when it gets to the word "select" for some reason, which is kind of a big problem. Hopefully I can fix that so I can finish it tonight.